test: extract MultiInstanceConcurrendModificationTest
The test needs processing without batching in order to reproduce concurrent modification of multi instance, otherwise it is not reproducible
